requirements.txt
python-telegram-bot==22.7
Flask==3.1.3
gunicorn==25.3.0